fix(deps): make this module compatible with typer 0.12
typer 0.12 no longer supports the [all] extras, they're now just part of the
default dependencies.  To get them for typer < 0.12 and yet be compatible with
0.12 too, all we need to do is add rich and shellingham explicitly ourselves.

The version specification I chose for rich and shellingham are those in the
current main branch of typer itself. These are very relaxed too, allowing
pretty much any version in the last 3 or 4 years.